+++ src/index.iife.ts
@@ -1,0 +1,17 @@
+export { MooColor as default } from './moo-color';
+export type {
+  AcceptedModel,
+  CmykColorString,
+  Color,
+  ColorData,
+  ColorInput,
+  ColorString,
+  HexColorString,
+  HexMode,
+  HslColorString,
+  HsvColorString,
+  HwbColorString,
+  RandomOptions,
+  RgbColorString,
+  RgbMode,
+} from './types';