+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- ESM migration
+  - `dist/index.js` Node.js (`import`)
+  - `dist/index.cjs` Node.js (`require`)
+  - `dist/index.d.ts`/`dist/index.d.cts` TypeScript types
+  - `dist/moo-color.global.js` minify for IIFE
+- update dependencies
+
 ## 1.0.3 - 2022-04-04
 
 - Lint codes.